+++ example/lib/lite_state_controllers/single_use_controller.dart
@@ -1,6 +1,7 @@
+import 'package:flutter/foundation.dart';
 import 'package:lite_state/lite_state.dart';
 
-SingleUseController get singleuseController {
+SingleUseController get singleUseController {
   return findController<SingleUseController>();
 }
 
@@ -8,6 +9,44 @@ class SingleUseController extends LiteStateController<SingleUseController> {
   int _counter = 0;
   int get counter {
     return _counter;
+  }
+
+  String? get date {
+    final date = getPersistentValue('date');
+    return date?.toString() ?? 'none';
+  }
+
+  Future setDate() async {
+    await setPersistentValue(
+      'date',
+      DateTime.now(),
+    );
+  }
+
+  Future setList(List value) async {
+    await setPersistentValue(
+      'list',
+      value,
+    );
+
+    final list = getPersistentValue('list');
+    if (kDebugMode) {
+      print(list);
+    }
+  }
+
+  Future saveMap() async {
+    await setPersistentValue(
+      'map',
+      {
+        'inMapList': ['one', 'two']
+      },
+    );
+
+    final map = getPersistentValue('map');
+    if (kDebugMode) {
+      print(map);
+    }
   }
 
   set counter(int value) {